backend/requirements.txt
# Web框架
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0,<3.0.0

# 数据库
aiosqlite>=0.19.0
# motor>=3.3.0,<4.0.0  # MongoDB 驱动（已注释，使用 SQLite）

# 认证安全
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
email-validator>=2.0.0

# 文件上传与解析
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=0.8.11
python-magic-bin>=0.4.14

# 网络请求与解析
httpx>=0.25.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 工具
python-dotenv>=1.0.0

# 可选：Redis缓存（分布式部署时使用）
# redis>=5.0.0
